v0.1 SBI replacement extensions
This patch tries to replace some of the required legacy extensions
with new extensions/functions that follows v0.2 calling convention
and avoids the issues that 0.1 extensions have. There are few new
functions calls such as hfence and reboot function calls which are
required for hypervisor use cases.

This patch is based on the HSM extension that is already sent as a PR.

Signed-off-by: Atish Patra <atish.patra@wdc.com>
Signed-off-by: Anup Patel <anup.patel@wdc.com>