Improve PMU extension with snapshot feature
Snapshot of PMU state allows the supervisor to read the counter values
or scountovf without any traps. This is mostly applicable for trap n emulate
scenarios in virtualization where the hypervisor traps for these CSRs.

In an non-virtualized world, the cost of CSR access vs dram access needs
to be considered.

Signed-off-by: Atish Patra <atishp@rivosinc.com>